Move HttpError alongside other error objects.
Keep types together.
