+++ app/middleware/404.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const HttpError = require('lib/wiring/http-error');
+const HttpError = require('lib/wiring/errors/http-error');
 
 const notFound = (request, response, next) => {
   next(new HttpError(404));
