+++ lib/wiring/errors/http-error.js
